解决服务商电站分布数据结构

+++ yn-core/src/main/java/com/yn/dao/StationDao.java
@@ -60,5 +60,5 @@ public interface StationDao extends JpaRepository<Station, Long>, JpaSpecificati
        
     @Query(value="SELECT t2.province_text AS provinceName, COUNT(*) AS stationNum FROM station t1 INNER JOIN province t2 "
        		+ "ON t1.province_id = t2.id where t1.del = 0 and t1.status = 1 and t1.server_id=?1 GROUP BY t1.province_id ORDER BY t2.id ASC",nativeQuery=true)
-    Object[] stationFenbuById(Integer id);
+    Object[] stationFenbuById(Long id);
 }
